Move ContentSource and FileContentSource into object module.
