+++ src/lib.rs
@@ -1,12 +1,6 @@
 mod attribution;
 pub use attribution::Attribution;
 
-mod content_source;
-pub use content_source::ContentSource;
-
-mod file_content_source;
-pub use file_content_source::FileContentSource;
-
 mod file_mode;
 pub use file_mode::FileMode;
