+++ src/object/content_source.rs
